refactor: move landpks slices and related code to landpks repo
